Reporte de credito

+++ modelo/punto_venta/limpiar-preventa.php
@@ -11,14 +11,19 @@ if (!isset($_SESSION['id_usuario'])) {
     header("Location:../../login.php");
 }
 
-$id_usuario = $_SESSION["id_usuario"];
-$sql = "DELETE FROM productos_preventa WHERE id_usuario = ?";
-$stmt = $con->prepare($sql);
-$stmt->bind_param('s', $id_usuario);
-$stmt->execute();
-$stmt->close();
+$id_usuario = isset($_SESSION["id_usuario"]) ? $_SESSION['id_usuario'] : '';
+if($id_usuario == ''){
+    responder(false, 'No existe la sesión, recarga la pagina', 'success', [], true, true);
+}else{
+    $sql = "DELETE FROM productos_preventa WHERE id_usuario = ?";
+    $stmt = $con->prepare($sql);
+    $stmt->bind_param('s', $id_usuario);
+    $stmt->execute();
+    $stmt->close();
+    
+    responder(true, 'Carrito limpiado con exito', 'success', [], true, true);
+}
 
-responder(true, 'Carrito limpiado con exito', 'success', [], true, true);
